requirements.txt
streamlit>=1.28.0
qrcode>=7.4.2
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
cryptography>=41.0.0